# src/main/resources/static/swagger.yaml
swagger: '2.0'
info:
  description: Sample Spring boot API
  version: 1.0.0
  title: Stock API
  contact:
    name: Zarin
    url: 'https://github.com/zarinlo'
host: 'localhost:8080'
basePath: /
tags:
  - name: stock-controller
    description: Stock Controller
paths:
  /api/v1/stocks:
    get:
      tags:
        - stock-controller
      summary: Get all stocks
      operationId: getAllStocksUsingGET
      consumes:
        - '*/*'
        - application/x-www-form-urlencoded
      produces:
        - application/json
      responses:
        '200':
          description: 'Successful: Stock(s) found.'
          schema:
            $ref: '#/definitions/StockGeneralResponse'
        '400':
          description: 'Bad Request: Check input parameter(s) syntax for invalid characters.'
        '401':
          description: 'Unauthorized: User is not entitled to retrieve information.'
        '404':
          description: 'Not Found: Stock(s) not found.'
        '500':
          description: 'Internal Server Error: Backend service is down.'
      deprecated: false
    post:
      tags:
        - stock-controller
      summary: Create a new stock
      operationId: createStockUsingPOST
      consumes:
        - application/json
        - '*/*'
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - in: body
          name: stock
          description: New stock object
          required: true
          schema:
            $ref: '#/definitions/Stock'
      responses:
        '200':
          description: 'Successful: Stock(s) found.'
          schema:
            $ref: '#/definitions/StockGeneralResponse'
        '400':
          description: 'Bad Request: Check input parameter(s) syntax for invalid characters.'
        '401':
          description: 'Unauthorized: User is not entitled to retrieve information.'
        '404':
          description: 'Not Found: Stock(s) not found.'
        '500':
          description: 'Internal Server Error: Backend service is down.'
      deprecated: false
  '/api/v1/stocks/{symbol}':
    get:
      tags:
        - stock-controller
      summary: Get a stock by symbol
      operationId: getStockBySymbolUsingGET
      consumes:
        - '*/*'
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: symbol
          in: path
          required: true
          type: string
          enum:
            - SIEMENS
            - INDUSTOWER
            - NIFTY NEXT 50
      responses:
        '200':
          description: 'Successful: Stock(s) found.'
          schema:
            $ref: '#/definitions/StockGeneralResponse'
        '400':
          description: 'Bad Request: Check input parameter(s) syntax for invalid characters.'
        '401':
          description: 'Unauthorized: User is not entitled to retrieve information.'
        '404':
          description: 'Not Found: Stock(s) not found.'
        '500':
          description: 'Internal Server Error: Backend service is down.'
      deprecated: false
    put:
      tags:
        - stock-controller
      summary: Update an existing stock by symbol
      operationId: updateStockBySymbolUsingPUT
      consumes:
        - application/json
        - '*/*'
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: lastPrice
          in: query
          description: Last Price
          required: true
          type: number
          format: double
          allowEmptyValue: false
          x-example: 127.05
        - name: symbol
          in: path
          description: A stock symbol
          required: true
          type: string
      responses:
        '200':
          description: 'Successful: Stock(s) found.'
          schema:
            $ref: '#/definitions/StockGeneralResponse'
        '400':
          description: 'Bad Request: Check input parameter(s) syntax for invalid characters.'
        '401':
          description: 'Unauthorized: User is not entitled to retrieve information.'
        '404':
          description: 'Not Found: Stock(s) not found.'
        '500':
          description: 'Internal Server Error: Backend service is down.'
      deprecated: false
    delete:
      tags:
        - stock-controller
      summary: Delete a stock by symbol
      operationId: deleteStockBySymbolUsingDELETE
      consumes:
        - '*/*'
        - application/x-www-form-urlencoded
      produces:
        - application/json
      parameters:
        - name: symbol
          in: path
          description: A stock symbol
          required: true
          type: string
      responses:
        '200':
          description: 'Successful: Stock(s) found.'
          schema:
            $ref: '#/definitions/StockGeneralResponse'
        '400':
          description: 'Bad Request: Check input parameter(s) syntax for invalid characters.'
        '401':
          description: 'Unauthorized: User is not entitled to retrieve information.'
        '404':
          description: 'Not Found: Stock(s) not found.'
        '500':
          description: 'Internal Server Error: Backend service is down.'
      deprecated: false
definitions:
  Stock:
    type: object
    properties:
      dayHigh:
        type: number
        format: double
      dayLow:
        type: number
        format: double
      identifier:
        type: string
      lastPrice:
        type: number
        format: double
      lastUpdateTime:
        type: string
      open:
        type: number
        format: double
      previousClose:
        type: number
        format: double
      symbol:
        type: string
      totalTradedValue:
        type: number
        format: double
      totalTradedVolume:
        type: integer
        format: int64
      yearHigh:
        type: integer
        format: int64
      yearLow:
        type: integer
        format: int64
    title: Stock
  StockGeneralResponse:
    type: object
    properties:
      response:
        type: object
      status:
        type: string
        enum:
          - 100 CONTINUE
          - 101 SWITCHING_PROTOCOLS
          - 102 PROCESSING
          - 103 CHECKPOINT
          - 200 OK
          - 201 CREATED
          - 202 ACCEPTED
          - 203 NON_AUTHORITATIVE_INFORMATION
          - 204 NO_CONTENT
          - 205 RESET_CONTENT
          - 206 PARTIAL_CONTENT
          - 207 MULTI_STATUS
          - 208 ALREADY_REPORTED
          - 226 IM_USED
          - 300 MULTIPLE_CHOICES
          - 301 MOVED_PERMANENTLY
          - 302 FOUND
          - 302 MOVED_TEMPORARILY
          - 303 SEE_OTHER
          - 304 NOT_MODIFIED
          - 305 USE_PROXY
          - 307 TEMPORARY_REDIRECT
          - 308 PERMANENT_REDIRECT
          - 400 BAD_REQUEST
          - 401 UNAUTHORIZED
          - 402 PAYMENT_REQUIRED
          - 403 FORBIDDEN
          - 404 NOT_FOUND
          - 405 METHOD_NOT_ALLOWED
          - 406 NOT_ACCEPTABLE
          - 407 PROXY_AUTHENTICATION_REQUIRED
          - 408 REQUEST_TIMEOUT
          - 409 CONFLICT
          - 410 GONE
          - 411 LENGTH_REQUIRED
          - 412 PRECONDITION_FAILED
          - 413 PAYLOAD_TOO_LARGE
          - 413 REQUEST_ENTITY_TOO_LARGE
          - 414 URI_TOO_LONG
          - 414 REQUEST_URI_TOO_LONG
          - 415 UNSUPPORTED_MEDIA_TYPE
          - 416 REQUESTED_RANGE_NOT_SATISFIABLE
          - 417 EXPECTATION_FAILED
          - 418 I_AM_A_TEAPOT
          - 419 INSUFFICIENT_SPACE_ON_RESOURCE
          - 420 METHOD_FAILURE
          - 421 DESTINATION_LOCKED
          - 422 UNPROCESSABLE_ENTITY
          - 423 LOCKED
          - 424 FAILED_DEPENDENCY
          - 425 TOO_EARLY
          - 426 UPGRADE_REQUIRED
          - 428 PRECONDITION_REQUIRED
          - 429 TOO_MANY_REQUESTS
          - 431 REQUEST_HEADER_FIELDS_TOO_LARGE
          - 451 UNAVAILABLE_FOR_LEGAL_REASONS
          - 500 INTERNAL_SERVER_ERROR
          - 501 NOT_IMPLEMENTED
          - 502 BAD_GATEWAY
          - 503 SERVICE_UNAVAILABLE
          - 504 GATEWAY_TIMEOUT
          - 505 HTTP_VERSION_NOT_SUPPORTED
          - 506 VARIANT_ALSO_NEGOTIATES
          - 507 INSUFFICIENT_STORAGE
          - 508 LOOP_DETECTED
          - 509 BANDWIDTH_LIMIT_EXCEEDED
          - 510 NOT_EXTENDED
          - 511 NETWORK_AUTHENTICATION_REQUIRED
      userMessages:
        type: array
        items:
          type: string
    title: StockGeneralResponse